+++ Source/Classes/Vehicle/Services/DeleteVehicleService.cs
@@ -6,6 +6,12 @@ public static class DeleteVehicleService
     var vehicle = GetVehicleByIdRepository.Get(id: id, connectionString: connectionString);
     if (vehicle == null) return Results.NotFound(new { message = "Veículo não encontrado." });
 
+    var policies = GetPolicyByClientRepository.Get(id: vehicle.id_cliente, connectionString: connectionString, pageNumber: 1, size: int.MaxValue);
+    if (policies.Any(policy => (policy.status == "Ativa" || policy.status == "Em Analise") && policy.id_veiculo == vehicle.id_veiculo)) return Results.BadRequest(new { message = "Não é possível desativar um veículo com apólices ativas ou em análise." });
+
+    var incidents = GetIncidentByClientRepository.Get(id: vehicle.id_cliente, connectionString: connectionString, pageNumber: 1, size: int.MaxValue);
+    if (incidents.Any(incident => (incident.status == "Andamento" || incident.status == "Processando") && incident.id_veiculo == vehicle.id_veiculo)) return Results.BadRequest(new { message = "Não é possível desativar um veículo com ocorrências ativas ou em processamento." });
+
     var result = DeleteVehicleRepository.Delete(id: id, connectionString: connectionString);
     if (result == 0) return Results.BadRequest(new { message = "Houve um erro ao processar sua requisição. Tente novamente mais tarde." });
